requirements.txt
aiogram==3.20.0.post0
python-dotenv==1.0.0
aiofiles==23.2.1
aiosqlite==0.19.0 